Ensures that entities that had tempKeys when they were exported are imported as having a tempKey.
This fixes a bug that caused the tempKey generator for numeric keys to duplicate keys when entities had been exported from, then imported back into an EntityManager.

When a new entity is attached to the EntityMangager, it has a temporary key generated, and the entityAspect.hasTempKey property is set to true.  However, when importing entities from the entityManager string serialization, the hasTempKey property was not being set, even when the entity was assigned a new tempKey.